Move attach_debugger function to debugging module
